example-python/pyproject.toml
[tool.poetry]
name = "example-python"
version = "0.1.0"
description = ""
authors = ["Jaroslaw Zabiello <hipertracker@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyyaml = "^6.0.1"
pyicu = "^2.12"
regex = "^2023.12.25"
rich-click = "^1.7.3"
alive-progress = "^3.1.5"
icecream = "^2.1.3"


[tool.poetry.group.dev.dependencies]
types-six = "^1.16.21.20240301"
types-requests = "^2.31.0.20240218"
types-pyyaml = "^6.0.12.12"
django-types = "^0.19.1"
ruff = "^0.3.0"
black = "^24.2.0"
types-regex = "^2023.12.25.20240106"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
